+++ src/features/filters/components/CommunityBuildFilters.tsx
@@ -1,0 +1,104 @@
+import { Archtype, ReleaseKey } from '@/features/items/types'
+import { useState } from 'react'
+import { DEFAULT_COMMUNITY_BUILD_FILTERS } from '@/features/filters/constants'
+import { CommunityBuildFilterProps } from '@/features/filters/types'
+import FiltersContainer from '@/features/filters/components/FiltersContainer'
+import ArchtypeFilters from '@/features/filters/components/ArchtypeFilters'
+import WeaponFilters from '@/features/filters/components/WeaponFilters'
+import ReleaseFilters from './ReleaseFilters'
+
+interface Props {
+  showBorder?: boolean
+  onUpdate: (filters: CommunityBuildFilterProps) => void
+}
+
+export default function CommunityBuildFilters({ onUpdate }: Props) {
+  const [filters, setFilters] = useState<CommunityBuildFilterProps>(
+    DEFAULT_COMMUNITY_BUILD_FILTERS,
+  )
+
+  function handleClearFilters() {
+    setFilters(DEFAULT_COMMUNITY_BUILD_FILTERS)
+    onUpdate(DEFAULT_COMMUNITY_BUILD_FILTERS)
+  }
+
+  function areAnyFiltersActive() {
+    return (
+      filters.archtypes.length > 0 ||
+      filters.longGun !== DEFAULT_COMMUNITY_BUILD_FILTERS['longGun'] ||
+      filters.handGun !== DEFAULT_COMMUNITY_BUILD_FILTERS['handGun'] ||
+      filters.melee !== DEFAULT_COMMUNITY_BUILD_FILTERS['melee'] ||
+      filters.selectedReleases.length < 2
+    )
+  }
+
+  function handleArchtypeChange(archtype: Archtype) {
+    let newArchtypes = [...filters.archtypes]
+
+    if (newArchtypes.includes(archtype)) {
+      newArchtypes = newArchtypes.filter((a) => a !== archtype)
+    } else {
+      // Only allow two archtypes to be selected at a time
+      if (filters.archtypes.length === 2) {
+        return
+      }
+      newArchtypes.push(archtype)
+    }
+
+    setFilters({
+      ...filters,
+      archtypes: newArchtypes,
+    })
+  }
+
+  function handleWeaponChange(
+    weapon: string,
+    type: 'longGun' | 'handGun' | 'melee',
+  ) {
+    setFilters({
+      ...filters,
+      [type]: weapon,
+    })
+  }
+
+  function handleReleaseChange(release: ReleaseKey) {
+    let newReleases = [...filters.selectedReleases]
+
+    if (newReleases.includes(release)) {
+      newReleases = newReleases.filter((r) => r !== release)
+    } else {
+      newReleases.push(release)
+    }
+
+    setFilters({
+      ...filters,
+      selectedReleases: newReleases,
+    })
+  }
+
+  return (
+    <FiltersContainer<CommunityBuildFilterProps>
+      areAnyFiltersActive={areAnyFiltersActive()}
+      filters={filters}
+      onApplyFilters={onUpdate}
+      onClearFilters={handleClearFilters}
+    >
+      <ArchtypeFilters
+        selectedArchtypes={filters.archtypes}
+        onChange={(archtype: Archtype) => handleArchtypeChange(archtype)}
+      />
+      <WeaponFilters
+        selectedLongGun={filters.longGun}
+        selectedHandGun={filters.handGun}
+        selectedMelee={filters.melee}
+        onChange={(weapon: string, type: 'longGun' | 'handGun' | 'melee') =>
+          handleWeaponChange(weapon, type)
+        }
+      />
+      <ReleaseFilters
+        selectedReleases={filters.selectedReleases}
+        onChange={(release: ReleaseKey) => handleReleaseChange(release)}
+      />
+    </FiltersContainer>
+  )
+}