+++ src/features/filters/constants.ts
@@ -1,0 +1,9 @@
+import { CommunityBuildFilterProps } from './types'
+
+export const DEFAULT_COMMUNITY_BUILD_FILTERS: CommunityBuildFilterProps = {
+  archtypes: [],
+  longGun: 'All',
+  handGun: 'All',
+  melee: 'All',
+  selectedReleases: ['base', 'dlc1'],
+}